% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timingRes.R
\name{build_cons_tree.timingRes}
\alias{build_cons_tree.timingRes}
\title{Build a majority rule consensus tree}
\usage{
\method{build_cons_tree}{timingRes}(x, compute_summaries = FALSE)
}
\arguments{
\item{x}{an object of class timingRes}

\item{compute_summaries}{should clade height quantiles be computed}
}
\value{
a list of: `phy` - the majority rule tree, `support` - clade supports, `ts_qs` - clade height quantiles
}
\description{
Build a majority rule consensus tree
}
