% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc_utils.R
\name{clade_node_counts.multiPhylo}
\alias{clade_node_counts.multiPhylo}
\title{Count the number of nodes in clades}
\usage{
\method{clade_node_counts}{multiPhylo}(x, clades)
}
\arguments{
\item{x}{an `ape::multiPhylo` object}

\item{clades}{a list of vectors of tip labels corresponding to clades}
}
\value{
a matrix of node counts. Rows correspond to phylogenies. Columns correspond to clades.
}
\description{
Count the number of nodes in clades
}
