% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc_utils.R
\name{clade_node_counts.phylo}
\alias{clade_node_counts.phylo}
\title{Count the number of nodes in clades}
\usage{
\method{clade_node_counts}{phylo}(x, clades)
}
\arguments{
\item{x}{an `ape::phylo` object}

\item{clades}{a list of vectors of tip labels corresponding to clades}
}
\value{
a vector of node counts for each clade in clades
}
\description{
Count the number of nodes in clades
}
