% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc_utils.R
\name{find_clade_index.multiPhylo}
\alias{find_clade_index.multiPhylo}
\title{Find the indices of the MRCAs of clades in a phylogeny}
\usage{
\method{find_clade_index}{multiPhylo}(x, clades)
}
\arguments{
\item{x}{an `ape::multiPhylo` object}

\item{clades}{a list of vectors of tip labels corresponding to clades}
}
\value{
a matrix of clade MRCA indices. Rows correspond to phylogenies. Columns correspond to clades.
}
\description{
Find the indices of the MRCAs of clades in a phylogeny
}
