% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc_utils.R
\name{find_clade_index.phylo}
\alias{find_clade_index.phylo}
\title{Find the indices of the MRCAs of clades in a phylogeny}
\usage{
\method{find_clade_index}{phylo}(x, clades)
}
\arguments{
\item{x}{an `ape::phylo` object}

\item{clades}{a list of vectors of tip labels corresponding to clades}
}
\value{
a vector of node indices correspoding to clade MRCAs
}
\description{
Find the indices of the MRCAs of clades in a phylogeny
}
