% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timingRes.R
\name{plot_densiCI.timingRes}
\alias{plot_densiCI.timingRes}
\title{Plot a modified densiTree}
\usage{
\method{plot_densiCI}{timingRes}(
  x,
  mrsd = NULL,
  n_samp = 100,
  layout = "rectangular",
  tip.order = NULL
)
}
\arguments{
\item{x}{an object of class timingRes}

\item{mrsd}{(Optional) most recent sampling time. Default: determine from tip times.}

\item{n_samp}{(Optional) number of draws to overlay in the densiTree. Default: 100}

\item{layout}{(Optional) tree layout passed to ggtree. Default: "rectangular"}

\item{tip.order}{(Optional) a vector specifying the order in which the tips should be displayed. Default: determine from data}
}
\description{
Plot a modified densiTree
}
