% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timingRes.R
\name{plot_mm_tree.multiPhylo}
\alias{plot_mm_tree.multiPhylo}
\title{Plot several multiple merger trees}
\usage{
\method{plot_mm_tree}{multiPhylo}(x, ...)
}
\arguments{
\item{x}{an object of class `ape::multiPhylo`}
}
\description{
Plot several multiple merger trees
}
