% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timingRes.R
\name{plot_mm_tree.phylo}
\alias{plot_mm_tree.phylo}
\title{Plot a multiple merger tree}
\usage{
\method{plot_mm_tree}{phylo}(x, ...)
}
\arguments{
\item{x}{an object of class `ape::phylo`}
}
\description{
Plot a multiple merger tree
}
