% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timingRes.R
\name{plot_pars.timingRes}
\alias{plot_pars.timingRes}
\title{Plot parameter pair marginals}
\usage{
\method{plot_pars}{timingRes}(x)
}
\arguments{
\item{x}{an object of class timingRes}
}
\description{
Plot parameter pair marginals
}
