% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timingRes.R
\name{plot_traces.timingRes}
\alias{plot_traces.timingRes}
\title{Plot parameter traces}
\usage{
\method{plot_traces}{timingRes}(x)
}
\arguments{
\item{x}{an object of class timingRes}
}
\description{
Plot parameter traces
}
