% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\docType{data}
\name{PPAR_DEFAULTS}
\alias{PPAR_DEFAULTS}
\title{Prior hyperparameter default values}
\format{
An object of class \code{list} of length 12.
}
\usage{
PPAR_DEFAULTS
}
\description{
Prior hyperparameter default values
}
\keyword{datasets}
