% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\docType{data}
\name{PPAR_NAMES}
\alias{PPAR_NAMES}
\title{Prior hyperparameter names}
\format{
An object of class \code{character} of length 10.
}
\usage{
PPAR_NAMES
}
\description{
Prior hyperparameter names
}
\keyword{datasets}
