% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timingRes.R
\name{sample_timetree.timingRes}
\alias{sample_timetree.timingRes}
\title{Sample time tree from the posterior}
\usage{
\method{sample_timetree}{timingRes}(x, n_samp = 1, replace = TRUE)
}
\arguments{
\item{x}{an object of class timingRes}

\item{n_samp}{number of samples to draw}

\item{replace}{sample with replacement}
}
\value{
an object of class `ape::phylo` or `ape::multiPhylo`
}
\description{
Sample time tree from the posterior
}
