% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_beta}
\alias{simulate_beta}
\title{Simulate a phylogeny from the beta coalescent}
\usage{
simulate_beta(samp_times, n_samp, nu, alpha)
}
\arguments{
\item{nu}{nu}

\item{alpha}{alpha^*}
}
\value{
a list consisting of sampling dates for each tip and the resulting phylogeny
}
\description{
Simulate a phylogeny from the beta coalescent
}
