% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_durret_schweinsberg}
\alias{simulate_durret_schweinsberg}
\title{Simulate a phylogeny from a particular durrett and schweinsberg coalescent. Experimental.}
\usage{
simulate_durret_schweinsberg(samp_times, n_samp, phi, nu)
}
\arguments{
\item{samp_times}{sampling epochs}

\item{n_samp}{number of samples per epoch}

\item{phi}{phi}

\item{nu}{nu}
}
\value{
a list consisting of sampling dates for each tip and the resulting phylogeny
}
\description{
Simulate a phylogeny from a particular durrett and schweinsberg coalescent. Experimental.
}
