% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_kingman}
\alias{simulate_kingman}
\title{Simulate a phylogeny from kingmans coalescent}
\usage{
simulate_kingman(samp_times, n_samp, nu)
}
\arguments{
\item{samp_times}{sampling epochs}

\item{n_samp}{number of samples per epoch}

\item{nu}{nu}
}
\value{
a list consisting of sampling dates for each tip and the resulting phylogeny
}
\description{
Simulate a phylogeny from kingmans coalescent
}
