% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_km_beta}
\alias{simulate_km_beta}
\title{Simulate a phylogeny from the extended beta coalescent}
\usage{
simulate_km_beta(samp_times, n_samp, phi, nu, alpha)
}
\arguments{
\item{samp_times}{sampling epochs}

\item{n_samp}{number of samples per epoch}

\item{phi}{phi}

\item{nu}{nu}

\item{alpha}{alpha^*}
}
\value{
a list consisting of sampling dates for each tip and the resulting phylogeny
}
\description{
Simulate a phylogeny from the extended beta coalescent
}
