% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_mut_arc}
\alias{simulate_mut_arc}
\title{Simulate a mutation scaled phylogeny from a coalescent genealogy}
\usage{
simulate_mut_arc(phy, mu, omega)
}
\arguments{
\item{phy}{genealogy}

\item{mu}{mu}

\item{omega}{omega}
}
\value{
a mutation scaled phylogeny
}
\description{
Simulate a mutation scaled phylogeny from a coalescent genealogy
}
